Int to Long conversion

+++ src/RealEstateInvesting.API/DTOs/Properties/UpdatePropertyBeforeActivationDto.cs
@@ -9,6 +9,6 @@ public class UpdatePropertyBeforeActivationDto
     public string PropertyType { get; set; } = default!;
     public string? ImageUrl { get; set; }
     public decimal InitialValuation { get; set; }
-    public int TotalUnits { get; set; }
+    public long TotalUnits { get; set; }
     public decimal AnnualYieldPercent { get; set; }
 }
