+++ src/RealEstateInvesting.Application/Admin/Properties/DTOs/ActivatePropertyDto.cs
@@ -2,7 +2,7 @@ namespace RealEstateInvesting.Application.Properties.Dtos;
 
 public class ActivatePropertyDto
 {
-    public int TotalUnits { get; set; }
+    public long TotalUnits { get; set; }
     public decimal RentalIncome { get; set; }
     public decimal AnnualYieldPercent { get; set; }
     public decimal RiskScore { get; set; }
